feat: add 'GetInviteCandidates' RPC and 'SearchQuery' message
This call should replace the 'GetAllUsers' call in the frontend used when suggesting users to invite into a project. As this could easily leak user data, the access to the entire user list should be restricted. Therefore, only matching suggestions are returned by the new call instead of all possible user.